+++ laktory/_testing/stockprices.py
@@ -1,6 +1,7 @@
 import os
 import pandas as pd
 from pyspark.sql import SparkSession
+import polars as pl
 
 spark = SparkSession.builder.appName("UnitTesting").getOrCreate()
 spark.conf.set("spark.sql.session.timeZone", "UTC")
@@ -11,6 +12,8 @@ spark.conf.set("spark.sql.session.timeZone", "UTC")
 # --------------------------------------------------------------------------- #
 
 data_dirpath = os.path.join(os.path.dirname(__file__), "../../tests/data/")
+
+# Spark
 df_brz = spark.read.parquet(os.path.join(data_dirpath, "brz_stock_prices"))
 df_slv = spark.read.parquet(os.path.join(data_dirpath, "slv_stock_prices"))
 df_meta = spark.read.parquet(os.path.join(data_dirpath, "slv_stock_meta"))
@@ -22,3 +25,11 @@ df_name = spark.createDataFrame(
         }
     )
 )
+
+# Polars
+dirpath = os.path.join(data_dirpath, "slv_stock_prices")
+for filename in os.listdir(dirpath):
+    if filename.endswith(".parquet"):
+        break
+df_slv_polars = pl.read_parquet(os.path.join(dirpath, filename))
+